app/requirements.txt
django==3.2
django-cors-headers==4.2.0
django-debug-toolbar==3.4.0
django-extensions==3.2.3
django-split-settings==1.2.0
elasticsearch==8.10.0
flake8==4.0
psycopg2-binary==2.9.7
pydantic==2.4.2
python-dotenv==1.0.0
uWSGI==2.0.22

#dev
#django-stubs
#ipython
#mypy
#isort

